celery07/requirements.txt
amqp==5.1.1
celery==5.2.7
ipython==8.5.0
redis==4.3.4
watchdog==2.1.9
librabbitmq==2.0.0
django-oscar==3.1
Django==3.2.15
requests==2.28.1
sorl_thumbnail==12.9.0
